+++ Assets/Resources/Brushes/Shared/Shaders/Special/IntersectionNonGeomFallback.shader
@@ -70,7 +70,7 @@ Shader "Brush/Special/Intersection" {
       half4 frag(v2f i) : COLOR
       {
         uint triangleIndex = uint(i.triangleids.x);
-        return PackUint16x2ToRgba8(uint2(_BatchID, triangleIndex));
+        return PackUint16x2ToRgba8(uint2((uint)_BatchID, triangleIndex));
       }
         ENDCG
     }
